requirements.txt
streamlit==1.21.0
openai==0.27.4
python-dotenv==1.0.0
duckduckgo-search==2.8.6
beautifulsoup4==4.12.2
tiktoken==0.3.3
